implements future torrent file comparing add this function to interface (mylist.php)

+++ mylist.php
@@ -7,8 +7,12 @@ $config = $c->getConfig();
 
 $rt = new RutrackerAPI();
 $rt->parse_user_params();
+// Items from tracker
 $flist = $rt->getFutureList();
+// Items from saved file
 $file = $rt->read_from_file();
+
+$items_from_file = array();
 
 ?>
 
@@ -101,21 +105,25 @@ $file = $rt->read_from_file();
             </h4>
         </div>
         <div class="col-md-6">
-            <?if(array_key_exists('error', $file)) {
+            <?if(count($file) > 1 && array_key_exists('error', $file)) {
                 echo '<div class="alert alert-info" role="alert">' . $file['error'] . "</div>";
-            }?>
+            }else{
+                echo '<div class="alert alert-info" role="alert">Successfully read from file</div>';
+                $items_from_file = $rt->compare_file_to_tracker_future_list($file, $flist);
+            }
+            ?>
         </div>
         <div class="col-md-2 button-div align-center">
             <label for="save-btn">Save list to file</label>
             <button type="button" id="save-btn" class="btn btn-default" value="Save to file">
-                <span class="glyphicon glyphicon-circle-arrow-down"></span>
+                <span class="glyphicon glyphicon glyphicon-floppy-save"></span>
                 Save
             </button>
         </div>
         <div class="col-md-2 button-div align-center">
             <label for="load-btn">Load list from file</label>
             <button type="button" id="load-btn" class="btn btn-default" value="Save to file">
-                <span class="glyphicon glyphicon-circle-arrow-up"></span>
+                <span class="glyphicon glyphicon glyphicon-floppy-open"></span>
                 Load
             </button>
         </div>
@@ -153,15 +161,42 @@ $file = $rt->read_from_file();
                 <?
                 for ($i = 0; $i < count($flist); $i++) {
                     echo "<tr>";
-                    echo '<td>' . ($i + 1 ) . '</td>';
+                    if(array_key_exists('type', $flist[$i])){
+                           if($flist[$i]['type'] == 'old'){
+                               echo '<td class="align-center" title="Saved in file and in tracker list">
+                                        <span class=" glyphicon glyphicon-asterisk color-orange"></span>
+                                    </td>';
+                           }elseif($flist[$i]['type'] == 'new'){
+                               echo '<td class="align-center" title="New item from tracker. Recommend save to file">
+                                        <span class="glyphicon glyphicon-flash color-green"></span>
+                                    </td>';
+                           }
+                    }else{
+                        echo '<td> - </td>';
+                    }
                     echo '<td><a href="' . $flist[$i]["topic_link"] . '">
                        ' . $flist[$i]['topic_name'] . ' </a></td>';
                     echo '<td><a href="' . $flist[$i]['link'] . '">
                     ' . $flist[$i]['name'] . '</a></td>';
-                    echo '<td class="color-green">' . $flist[$i]['seeds'] . '</td>';
-                    echo '<td class="color-red">' . $flist[$i]['leeches'] . '</td>';
+                    echo '<td class="color-green align-center">' . $flist[$i]['seeds'] . '</td>';
+                    echo '<td class="color-red align-center">' . $flist[$i]['leeches'] . '</td>';
                     echo "</tr>";
-                } ?>
+                }
+                if(count($items_from_file) > 0) {
+                    foreach ($items_from_file as $item) {
+                        echo "<tr>";
+                        echo "<td class='align-center'
+                              title='Saved only in file. Probably removed from tracker list'>
+                            <span class='glyphicon glyphicon-floppy-disk col-more-gray'></span>
+                            </td>";
+                        echo "<td class='align-center'> - </td>";
+                        echo "<td><a href='".$item['link'] . "'>" . $item['name'] . "</a></td>";
+                        echo "<td class='align-center cursor-pointer' title='Get seeders count'> ? </td>";
+                        echo "<td class='align-center cursor-pointer' title='Get leechers count'> ? </td>";
+                        echo "</tr>";
+                    }
+                }
+                ?>
 
             </tbody>
         </table>
